Simplify the Devenv program definitions
The Devenv objecvt is the actual "program", and this objec ynow
contains nothing but the exposed program operations.

The Devenv operations now use EitherT to handle "early-exit", to avoid
the problem of needing nested if statements and for
comprehensions. This is a trade-off between the complexity of using
EitherT behind the scenes vs the clarity of the program's execution
flow.